[HWKINVENT-20] Add impl-tinkerpop-spi to be able to configure graphs in a vendor-specific manner.
This is because some databases (notably Titan) cannot be configured fully
using strictly only Blueprints API.
