+++ impl-tinkerpop-parent/impl/src/main/java/org/hawkular/inventory/impl/tinkerpop/FilterVisitor.java
