+++ impl-tinkerpop-parent/impl/src/main/resources/META-INF/services/org.hawkular.inventory.api.Inventory
