+++ impl-tinkerpop-parent/pom.xml
@@ -20,25 +20,26 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <parent>
+    <artifactId>inventory-parent</artifactId>
+    <groupId>org.hawkular.inventory</groupId>
+    <version>0.0.1-SNAPSHOT</version>
+  </parent>
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>org.hawkular.inventory</groupId>
-  <artifactId>inventory-parent</artifactId>
-  <version>0.0.1-SNAPSHOT</version>
+  <artifactId>inventory-impl-tinkerpop-parent</artifactId>
   <packaging>pom</packaging>
 
-  <inceptionYear>2015</inceptionYear>
+  <name>Hawkular Inventory Tinkerpop implementation parent POM</name>
 
-  <parent>
-    <groupId>org.hawkular</groupId>
-    <artifactId>hawkular-parent</artifactId>
-    <version>9</version>
-  </parent>
+  <properties>
+    <version.com.tinkerpop>2.6.0</version.com.tinkerpop>
+  </properties>
 
   <modules>
-    <module>api</module>
-    <module>impl-tinkerpop-parent</module>
-    <module>rest-servlet</module>
-    <module>rest-test</module>
+    <module>impl</module>
+    <module>spi</module>
+    <module>tinkergraph-provider</module>
+    <module>titan-provider</module>
   </modules>
 </project>
