+++ impl-tinkerpop-parent/tinkergraph-provider/src/main/resources/META-INF/services/org.hawkular.inventory.impl.tinkerpop.spi.GraphProvider
@@ -1,0 +1,1 @@
+org.hawkular.inventory.impl.tinkerpop.provider.TinkerGraphProvider
