+++ impl-tinkerpop-parent/titan-provider/src/main/resources/META-INF/services/org.hawkular.inventory.impl.tinkerpop.spi.GraphProvider
@@ -1,0 +1,1 @@
+org.hawkular.inventory.impl.tinkerpop.provider.TitanProvider
